Migrate Sankey renderer to React
